news-platform-final/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.14.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
python-multipart>=0.0.9
celery[redis]>=5.4.0
redis>=5.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
feedparser>=6.0.11
newspaper3k>=0.2.8
langdetect>=1.0.9
deep-translator>=1.11.4
anthropic>=0.39.0
openai>=1.50.0
groq>=0.9.0
google-genai>=1.0.0
python-dotenv>=1.0.1
tenacity>=8.4.0
apscheduler>=3.10.4
aiohttp>=3.10.0
youtube-transcript-api>=0.6.0
lxml_html_clean
requests>=2.31.0
